Add files via upload
SO Where am I? Great Question If only i Knew 

Just kidding heres the scoop. I am trying to implement the % operator in order to make sure every number is unique and not the same. This is proving to be a bitch. However I am sure this is the last hump before I am done.